+++ src/main/java/mildthreat/framecalc/service/FrameCalcServiceImpl.java
@@ -3,11 +3,15 @@ package mildthreat.framecalc.service;
 import mildthreat.framecalc.data.payloads.request.FrameCalcRequest;
 import mildthreat.framecalc.data.payloads.response.FrameCalcResponse;
 import mildthreat.framecalc.data.payloads.response.TelemetryResponse;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
+import org.springframework.security.oauth2.core.OAuth2AccessToken;
 import org.springframework.stereotype.Service;
 
 @Service
 public class FrameCalcServiceImpl implements FrameCalcService {
-    private static int usageCount = 0;
+    private static long usageCount = 0;
 
     @Override
     public FrameCalcResponse calculateFrameDetails(FrameCalcRequest frameCalcRequest) {
@@ -22,6 +26,5 @@ public class FrameCalcServiceImpl implements FrameCalcService {
     public TelemetryResponse getTelemetryResults() {
         return new TelemetryResponse(usageCount);
     }
-
 
 }
